service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlmodel
openai
pymupdf
pypdf
debugpy
alembic
Pillow